Caronas e Ajustes em banco
Adicionada a camada de negócio de Caronas, atualmente cadastra uma solicitação e lista as solicitações criadas à partir de filtros enviados.

Criado um serviço específico para caronas (o plano é deixar tudo separado de forma que seja fácil a manutenção)

No banco de dados eu ajustei os tipos usados em latitude e longitude para float, podendo usar double no c# (pesquisei e vi que era melhor)

Criei método que lista os amigos do usuário

+++ ElmatSvc/RIDE.cs
@@ -17,10 +17,10 @@ namespace ElmatSvc
         public int RideID { get; set; }
         public int UserID { get; set; }
         public System.DateTime Hour { get; set; }
-        public long LatOrg { get; set; }
-        public long LonOrg { get; set; }
-        public long LatDest { get; set; }
-        public long LonDest { get; set; }
+        public double LatOrg { get; set; }
+        public double LonOrg { get; set; }
+        public double LatDest { get; set; }
+        public double LonDest { get; set; }
     
         public virtual USER USER { get; set; }
     }
